requirements.txt
# AgentForge Python Dependencies
# Production-ready configuration

# ===========================================
# Web Framework
# ===========================================
FastAPI>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ===========================================
# Database
# ===========================================
asyncpg>=0.29.0
sqlalchemy>=2.0.0
redis>=5.0.0

# ===========================================
# Vector Database
# ===========================================
qdrant-client>=1.7.0

# ===========================================
# HTTP Client
# ===========================================
httpx>=0.25.0
aiohttp>=3.9.0

# ===========================================
# Configuration
# ===========================================
python-dotenv>=1.0.0
pyyaml>=6.0

# ===========================================
# Logging
# ===========================================
loguru>=0.7.0

# ===========================================
# Security
# ===========================================
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# ===========================================
# Monitoring (Optional)
# ===========================================
prometheus-client>=0.19.0

# ===========================================
# Testing
# ===========================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ===========================================
# Code Quality
# ===========================================
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0

# ===========================================
# Type Stubs
# ===========================================
types-pyyaml>=6.0.0
types-redis>=4.6.0

# ===========================================
# Utilities
# ===========================================
python-multipart>=0.0.6
aiofiles>=23.2.0
tenacity>=8.2.0
croniter>=2.0.0
